embedding-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
